LayoutTests/imported/w3c/IndexedDB-private-browsing/idbfactory_open5-expected.txt


PASS IDBFactory.open() - new database is empty
